Sf 6 commithook

+++ contracts/libraries/SafeTransfer.sol
@@ -3,7 +3,6 @@ pragma solidity 0.6.12;
 import "../interfaces/IWETH9.sol";
 
 abstract contract SafeTransfer {
-
     address public immutable WETH9;
 
     constructor(address _WETH9) public {
@@ -11,13 +10,13 @@ abstract contract SafeTransfer {
     }
 
     receive() external payable {
-        require(msg.sender == WETH9, 'Not WETH9');
+        require(msg.sender == WETH9, "Not WETH9");
     }
 
     function _depositAssets(
         address _token,
-        address _payer, 
-        address _receiver, 
+        address _payer,
+        address _receiver,
         uint256 _amount
     ) internal {
         if (address(_token) == WETH9 && address(this).balance >= _amount) {
@@ -30,8 +29,8 @@ abstract contract SafeTransfer {
     }
 
     function _withdrawAssets(
-        address _token, 
-        address _receiver, 
+        address _token,
+        address _receiver,
         uint256 _amount
     ) internal {
         if (address(_token) == WETH9) {
@@ -50,7 +49,7 @@ abstract contract SafeTransfer {
 
     function _unwrapWETH(address _receiver, uint256 _amount) internal {
         uint256 balanceWETH9 = IWETH9(WETH9).balanceOf(address(this));
-        require(balanceWETH9 >= _amount, 'Insufficient WETH9');
+        require(balanceWETH9 >= _amount, "Insufficient WETH9");
 
         if (balanceWETH9 > 0) {
             IWETH9(WETH9).withdraw(_amount);
@@ -59,10 +58,19 @@ abstract contract SafeTransfer {
     }
 
     /// @dev Transfer helper from UniswapV2 Router
-    function _safeApprove(address token, address to, uint value) internal {
+    function _safeApprove(
+        address token,
+        address to,
+        uint256 value
+    ) internal {
         // bytes4(keccak256(bytes('approve(address,uint256)')));
-        (bool success, bytes memory data) = token.call(abi.encodeWithSelector(0x095ea7b3, to, value));
-        require(success && (data.length == 0 || abi.decode(data, (bool))), 'TransferHelper: APPROVE_FAILED');
+        (bool success, bytes memory data) = token.call(
+            abi.encodeWithSelector(0x095ea7b3, to, value)
+        );
+        require(
+            success && (data.length == 0 || abi.decode(data, (bool))),
+            "TransferHelper: APPROVE_FAILED"
+        );
     }
 
     /**
@@ -76,12 +84,14 @@ abstract contract SafeTransfer {
         uint256 amount
     ) internal virtual {
         // solium-disable-next-line security/no-low-level-calls
-        (bool success, bytes memory data) =
-            token.call(
-                // 0xa9059cbb = bytes4(keccak256("transfer(address,uint256)"))
-                abi.encodeWithSelector(0xa9059cbb, to, amount)
-            );
-        require(success && (data.length == 0 || abi.decode(data, (bool))), 'TransferHelper: TRANSFER_FROM_FAILED'); // ERC20 Transfer failed
+        (bool success, bytes memory data) = token.call(
+            // 0xa9059cbb = bytes4(keccak256("transfer(address,uint256)"))
+            abi.encodeWithSelector(0xa9059cbb, to, amount)
+        );
+        require(
+            success && (data.length == 0 || abi.decode(data, (bool))),
+            "TransferHelper: TRANSFER_FROM_FAILED"
+        ); // ERC20 Transfer failed
     }
 
     function _safeTransferFrom(
@@ -90,23 +100,34 @@ abstract contract SafeTransfer {
         uint256 amount
     ) internal virtual {
         // solium-disable-next-line security/no-low-level-calls
-        (bool success, bytes memory data) =
-            token.call(
-                // 0x23b872dd = bytes4(keccak256("transferFrom(address,address,uint256)"))
-                abi.encodeWithSelector(0x23b872dd, from, address(this), amount)
-            );
-        require(success && (data.length == 0 || abi.decode(data, (bool))), 'TransferHelper: TRANSFER_FROM_FAILED'); // ERC20 TransferFrom failed
+        (bool success, bytes memory data) = token.call(
+            // 0x23b872dd = bytes4(keccak256("transferFrom(address,address,uint256)"))
+            abi.encodeWithSelector(0x23b872dd, from, address(this), amount)
+        );
+        require(
+            success && (data.length == 0 || abi.decode(data, (bool))),
+            "TransferHelper: TRANSFER_FROM_FAILED"
+        ); // ERC20 TransferFrom failed
     }
 
-    function _safeTransferFrom(address token, address from, address to, uint value) internal {
+    function _safeTransferFrom(
+        address token,
+        address from,
+        address to,
+        uint256 value
+    ) internal {
         // bytes4(keccak256(bytes('transferFrom(address,address,uint256)')));
-        (bool success, bytes memory data) = token.call(abi.encodeWithSelector(0x23b872dd, from, to, value));
-        require(success && (data.length == 0 || abi.decode(data, (bool))), 'TransferHelper: TRANSFER_FROM_FAILED');
+        (bool success, bytes memory data) = token.call(
+            abi.encodeWithSelector(0x23b872dd, from, to, value)
+        );
+        require(
+            success && (data.length == 0 || abi.decode(data, (bool))),
+            "TransferHelper: TRANSFER_FROM_FAILED"
+        );
     }
 
-    function _safeTransferETH(address to, uint value) internal {
-        (bool success,) = to.call{value:value}(new bytes(0));
-        require(success, 'TransferHelper: ETH_TRANSFER_FAILED');
+    function _safeTransferETH(address to, uint256 value) internal {
+        (bool success, ) = to.call{value: value}(new bytes(0));
+        require(success, "TransferHelper: ETH_TRANSFER_FAILED");
     }
-
 }
